requirements.txt
Pillow==11.0.0
streamlit==1.41.1
ultralytics==8.3.49
